+++ src/test/java/nextstep/subway/path/domain/PathFinderTest.java
@@ -71,7 +71,7 @@ class PathFinderTest {
     @DisplayName("최단거리를 조회한다")
     void findPathsDistanceTest() {
         // when
-        double pathsDistance = pathFinder.getPathsDistance(남부터미널역, 역삼역);
+        int pathsDistance = pathFinder.getPathsDistance(남부터미널역, 역삼역);
 
         // then
         assertThat(pathsDistance).isEqualTo(14);
